+++ clients/admin-ui/src/features/system/form.ts
@@ -59,6 +59,19 @@ export const transformSystemToFormValues = (
     },
     customFieldValues,
     privacy_policy: system.privacy_policy ? system.privacy_policy : "",
+    data_security_practices: system.data_security_practices
+      ? system.data_security_practices
+      : "",
+    // these fields require membership in their enums and won't let you assign them a blank string normally
+    // they're transformed back into appropriate systems on submission by transformFormValuesToSystem below
+    // @ts-ignore
+    legal_basis_for_profiling: system.legal_basis_for_profiling
+      ? system.legal_basis_for_profiling
+      : "",
+    // @ts-ignore
+    legal_basis_for_transfers: system.legal_basis_for_transfers
+      ? system.legal_basis_for_transfers
+      : "",
     data_stewards: dataStewards,
   };
 };